adjusted relative imports; updated CODEOWNERS
