+++ .github/CODEOWNERS
@@ -1,14 +1,15 @@
 # detection-rules code owners
 # POC: Elastic Security Intelligence and Analytics Team
 
-tests/**/*.py     @mikaayenson @eric-forte-elastic @terrancedejesus 
-detection_rules/  @mikaayenson @eric-forte-elastic @terrancedejesus 
-tests/            @mikaayenson @eric-forte-elastic @terrancedejesus 
-lib/              @mikaayenson @eric-forte-elastic @terrancedejesus 
-rta/              @mikaayenson @eric-forte-elastic @terrancedejesus 
+tests/**/*.py     @mikaayenson @eric-forte-elastic @terrancedejesus
+detection_rules/  @mikaayenson @eric-forte-elastic @terrancedejesus
+tests/            @mikaayenson @eric-forte-elastic @terrancedejesus
+lib/              @mikaayenson @eric-forte-elastic @terrancedejesus
+rta/              @mikaayenson @eric-forte-elastic @terrancedejesus
+hunting/          @mikaayenson @eric-forte-elastic @terrancedejesus
 
 # skip rta-mapping to avoid the spam
-detection_rules/etc/packages.yaml @mikaayenson @eric-forte-elastic @terrancedejesus 
-detection_rules/etc/*.json        @mikaayenson @eric-forte-elastic @terrancedejesus 
-detection_rules/etc/*.json        @mikaayenson @eric-forte-elastic @terrancedejesus 
-detection_rules/etc/*/*           @mikaayenson @eric-forte-elastic @terrancedejesus 
+detection_rules/etc/packages.yaml @mikaayenson @eric-forte-elastic @terrancedejesus
+detection_rules/etc/*.json        @mikaayenson @eric-forte-elastic @terrancedejesus
+detection_rules/etc/*.json        @mikaayenson @eric-forte-elastic @terrancedejesus
+detection_rules/etc/*/*           @mikaayenson @eric-forte-elastic @terrancedejesus
