+++ .github/paths-labeller.yml
@@ -14,6 +14,8 @@
   - "kql/**/*.py"
 - "RTA":
   - "rta/**/*"
+- "Hunting":
+  - "hunting/**/*"
 
 # rules
 - "bbr":
