+++ hunting/utils.py
@@ -3,6 +3,7 @@
 # 2.0; you may not use this file except in compliance with the Elastic License
 # 2.0.
 
+import tomllib
 from pathlib import Path
 
 import click
@@ -10,10 +11,10 @@ import urllib3
 import yaml
 
 from detection_rules.attack import tactics_map, technique_lookup
-from hunting.definitions import HUNTING_DIR
+from .definitions import HUNTING_DIR, Hunt
 
 
-def load_index() -> dict:
+def load_index_file() -> dict:
     """Load the hunting index.yml file."""
     index_file = HUNTING_DIR / "index.yml"
     if not index_file.exists():
@@ -26,66 +27,33 @@ def load_index() -> dict:
     return hunting_index
 
 
-def search_index(base_path: Path, mitre_filter: tuple = (), data_source: str = None) -> list:
-    """Search the index for queries matching the MITRE techniques, tactic, or data_source."""
+def load_toml(file_path: Path) -> Hunt:
+    """Load and validate TOML content as Hunt dataclass."""
+    if not file_path.is_file():
+        raise FileNotFoundError(f"TOML file not found: {file_path}")
 
-    # Load index.yml
+    contents = file_path.read_text(encoding="utf-8")
+    toml_dict = tomllib.loads(contents)
+
+    # Validate and load the content into the Hunt dataclass
+    return Hunt(**toml_dict["hunt"])
+
+
+def load_all_toml(base_path: Path):
+    """Load all TOML files from the directory and return a list of Hunt configurations and their paths."""
+    hunts = []
+    for toml_file in base_path.rglob("*.toml"):
+        hunt_config = load_toml(toml_file)
+        hunts.append((hunt_config, toml_file))
+    return hunts
+
+
+def save_index_file(base_path: Path, directories: dict) -> None:
+    """Save the updated index.yml file."""
     index_file = base_path / "index.yml"
-    if not index_file.exists():
-        click.echo(f"No index.yml found at {index_file}.")
-        return []
-
-    with open(index_file, 'r') as f:
-        hunting_index = yaml.safe_load(f)
-
-    # Initialize a set for MITRE technique IDs
-    mitre_technique_ids = set()
-    reverse_tactics_map = {v: k for k, v in tactics_map.items()}
-
-    # Process the MITRE filter (could be tactic ID, technique ID, or sub-technique ID)
-    for filter_item in mitre_filter:
-        if filter_item in reverse_tactics_map:
-            # Retrieve the tactic name using the tactic ID
-            tactic_name = reverse_tactics_map[filter_item]
-            click.echo(f"Found tactic ID {filter_item} (Tactic Name: {tactic_name}). Searching for associated techniques.")  # noqa: E501
-
-            # Now find techniques that have this tactic in the kill_chain_phases
-            for tech_id, details in technique_lookup.items():
-                kill_chain_phases = details.get('kill_chain_phases', [])
-                # Match based on phase_name, ensuring it's lowercased and without hyphens
-                if any(tactic_name.lower().replace(' ', '-') == phase['phase_name'] for phase in kill_chain_phases):
-                    mitre_technique_ids.add(tech_id)
-
-        elif filter_item in technique_lookup:
-            # Add the technique or sub-technique ID directly
-            mitre_technique_ids.add(filter_item)
-
-            # Include all sub-techniques if the filter is a parent technique (e.g., T1078)
-            if '.' not in filter_item:
-                # Find all sub-techniques of the parent technique
-                sub_techniques = {
-                    sub_tech_id for sub_tech_id in technique_lookup
-                    if sub_tech_id.startswith(f"{filter_item}.")
-                }
-                mitre_technique_ids.update(sub_techniques)
-
-    # Search the index for queries that match the MITRE techniques and data_source
-    results = []
-    for folder, queries in hunting_index.items():
-        # If a data_source is specified, filter by data_source
-        if data_source and folder != data_source:
-            continue
-
-        for uuid, query in queries.items():  # Adjust to iterate over the dictionary
-            query_techniques = query.get('mitre', [])
-            # Match queries that contain at least one technique from the filtered set
-            if not mitre_technique_ids or any(tech in mitre_technique_ids for tech in query_techniques):
-                # Add the data_source (which is the folder or top-level object) to each result
-                query_with_data_source = query.copy()
-                query_with_data_source['data_source'] = folder  # Add data_source to the query result
-                results.append(query_with_data_source)
-
-    return results
+    with open(index_file, 'w') as f:
+        yaml.safe_dump(directories, f, default_flow_style=False, sort_keys=False)
+    print(f"Index YAML updated at: {index_file}")
 
 
 def validate_link(link: str):
@@ -94,3 +62,37 @@ def validate_link(link: str):
     response = http.request('GET', link)
     if response.status != 200:
         raise ValueError(f"Invalid link: {link}")
+
+
+def update_index_yml(base_path: Path) -> None:
+    """Update index.yml based on the current TOML files."""
+    directories = load_index_file()  # Load the existing index.yml data
+
+    # Load all TOML files recursively
+    toml_files = base_path.rglob("queries/*.toml")  # Find all TOML files in the 'queries' directory
+
+    for toml_file in toml_files:
+        # Load TOML and extract hunt configuration
+        hunt_config = load_toml(toml_file)  # Parse the TOML file
+
+        folder_name = toml_file.parent.parent.name  # Determine the folder (platform, integration, etc.)
+        uuid = hunt_config.uuid  # Use the UUID as the key
+
+        entry = {
+            'name': hunt_config.name,
+            'path': f"./{toml_file.relative_to(base_path).as_posix()}",  # Ensure the path links to TOML file
+            'mitre': hunt_config.mitre
+        }
+
+        # Check if the folder_name exists and if it's a list, convert it to a dictionary
+        if folder_name not in directories:
+            directories[folder_name] = {uuid: entry}  # Initialize as a dictionary
+        else:
+            if isinstance(directories[folder_name], list):
+                # Convert the list to a dictionary, using UUIDs as keys
+                directories[folder_name] = {item['uuid']: item for item in directories[folder_name]}
+            # Now we can safely use UUID as the key
+            directories[folder_name][uuid] = entry
+
+    # Save the updated index.yml
+    save_index_file(base_path, directories)
