+++ pyproject.toml
@@ -52,7 +52,7 @@ dev = ["pep8-naming==0.13.0", "PyGithub==2.2.0", "flake8==7.0.0", "pyflakes==3.2
 
 [tool.setuptools]
 package-data = {"kql" = ["*.g"]}
-packages = ["detection_rules", "rta"]
+packages = ["detection_rules", "rta", "hunting"]
 
 [tool.pytest.ini_options]
 filterwarnings = [
